+++ software/components/ible/include/ible/specification.hpp
@@ -46,7 +46,7 @@ class characteristic_t : public visitable_t<visitor_t<characteristic_t>>, public
   };
 
   characteristic_t(characteristic_t::id_t, bool, characteristic_t::property_t, characteristic_t::permission_t, uint8_t*,
-                   size_t length, size_t max_length);
+                   uint16_t length, uint16_t max_length);
   [[nodiscard]] std::string stringify() const override;
   friend class attribute_visitor;
 
@@ -56,7 +56,7 @@ class characteristic_t : public visitable_t<visitor_t<characteristic_t>>, public
   characteristic_t::property_t m_property;
   characteristic_t::permission_t m_permission;
   std::uint8_t* m_value;
-  size_t m_length, m_max_length;
+  uint16_t m_length, m_max_length;
 
   bool automated();
